generating test.js
